move models and add very basic tests
