+++ lib/database/database.dart
@@ -1,9 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:runningapp/models/run.dart';
-import 'package:runningapp/pages/logged_in/routes_page/route_model.dart';
-import 'package:runningapp/pages/logged_in/story_page/models/progress_model.dart';
-import 'package:runningapp/pages/logged_in/story_page/models/quests_model.dart';
+import 'package:runningapp/models/route_model.dart';
+import 'package:runningapp/models/progress_model.dart';
+import 'package:runningapp/models/quests_model.dart';
 import 'package:runningapp/pages/logged_in/story_page/models/story_model.dart';
 import 'package:runningapp/state/backend/authenticator.dart';
