+++ lib/pages/logged_in/run_page/components/run_detail_and_stop.dart
@@ -11,7 +11,7 @@ import 'package:runningapp/pages/logged_in/run_page/map_and_location_logic/draw_
 import 'package:runningapp/pages/logged_in/run_page/map_and_location_logic/google_maps_container.dart';
 import 'package:runningapp/pages/logged_in/run_page/map_and_location_logic/location_service.dart';
 import 'package:runningapp/pages/logged_in/run_page/paused_page/paused_page.dart';
-import 'package:runningapp/pages/logged_in/story_page/models/progress_model.dart';
+import 'package:runningapp/models/progress_model.dart';
 import 'package:runningapp/providers.dart';
 import 'package:stop_watch_timer/stop_watch_timer.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
